+++ src/test/unit/taglib_completion_logic.test.ts
@@ -9,7 +9,9 @@ const tags: ProjectTagLibTag[] = [
 		method: 'accountLink',
 		attributes: ['href', 'class'],
 		usesBody: false,
-		sourcePath: 'DemoUITagLib.groovy'
+		sourcePath: 'DemoUITagLib.groovy',
+		methodLine: 3,
+		methodColumn: 8
 	},
 	{
 		name: 'demoUI:adminButton',
@@ -17,7 +19,9 @@ const tags: ProjectTagLibTag[] = [
 		method: 'adminButton',
 		attributes: ['url', 'class'],
 		usesBody: true,
-		sourcePath: 'DemoUITagLib.groovy'
+		sourcePath: 'DemoUITagLib.groovy',
+		methodLine: 10,
+		methodColumn: 8
 	}
 ];